- Fixed EVENT_Types enum - Fixed minor warnings - Added varidic template arguments to ThreadPool jobs - Added exception propogation to ThreadPool - Added shutdown to ThreadPool

+++ Ghostbait/Ghostbait/Source Files/MeshManager.cpp
@@ -106,7 +106,7 @@ int MeshManager::AddElement(const char* _meshFilePath, const unsigned int _id = 
 {
 	meshPositionColor* mesh = GetElement(_id);
 	smallMeshes.push_back(*mesh);
-	return smallMeshes.size();
+	return (int)smallMeshes.size();
 }
